requirements.txt
Flask>=3.0
GitPython>=3.1
watchdog>=4.0
rich>=13.7
